Deployment/requirements.txt
streamlit
pandas
seaborn
numpy
matplotlib
scipy
scikit-learn==1.2.2
Pillow
plotly